requirements.txt
fastapi
uvicorn
pdfminer.six
python-docx
networkx
pillow
pytesseract
google-generativeai
